backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
pydantic==2.7.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9